feat: model과 component 파일 위치 수정
